로직 최적화

+++ src/com/marblegame/core/RuleEngine.java
@@ -176,7 +176,7 @@ public class RuleEngine {
             if (tile instanceof TouristSpot) {
                 TouristSpot spot = (TouristSpot) tile;
                 // 해당 플레이어가 잠근 관광지만 해제
-                if (spot.isLocked() && spot.lockedBy != null && spot.lockedBy == playerIndex) {
+                if (spot.isLocked() && spot.lockedBy != null && spot.lockedBy.intValue() == playerIndex) {
                     spot.setLocked(false);
                     spot.lockedBy = null;
                 }
